requirements.txt
spotipy==2.23.0
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
tabulate==0.9.0
flask==3.0.0
requests==2.31.0
